% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm-movement.R
\name{boundaries}
\alias{boundaries}
\title{Spatial boundary restrictions}
\usage{
boundaries(x, ...)
}
\arguments{
\item{x}{The positions of the particles.}

\item{\dots}{Additional arguments for different methods.}
}
\description{
Set spatial restrictions to the domain.
}
\details{
Boundaries is a generic and methods can be written. The default
applies simmetric boundaries (dynamics over a torus) or reflexive barriers.
}

