% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm-movement.R
\name{diffusion}
\alias{diffusion}
\title{Brownian diffusion of a set of particles}
\usage{
diffusion(object, sd, ...)
}
\arguments{
\item{object}{The positions of the particles, dimension is taken from the number
of columns or assumed to be 1 is no columns.}

\item{sd}{Standard deviation for the gaussian jump, for dynamics models should be
set proportional to \code{sqrt(dt)}.}

\item{\dots}{Additional arguments for different methods.}
}
\description{
This funtions performs a brownian difussion over a set of particles.
The dimension is automatically calculated from the number of columns of the object.
}
\details{
This functions apply a brownian diffusion to a set of point coordinates.
}

