% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibm-main.R
\docType{package}
\name{ibm-package}
\alias{ibm}
\alias{ibm-package}
\title{Individual Based Models in R}
\description{
Individual Based Models in R
}
\details{
\tabular{ll}{ Package: \tab ibm\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2015-04-27\cr License: \tab GPL-2\cr } ibm()
}
\examples{
LV.ibm()
}
\author{
Ricardo Oliveros-Ramos Maintainer: Ricardo Oliveros-Ramos
<ricardo.oliveros@gmail.com>
}
\references{
ibm
}
\keyword{ibm}

