% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm-mortality.R
\name{mortality}
\alias{mortality}
\title{Mortality Process}
\usage{
mortality(object, rates, ...)
}
\arguments{
\item{object}{The population object, containing the information about individuals.}

\item{rates}{The mortality rate or rates.}

\item{\dots}{Additional arguments for different methods.}
}
\description{
This functions performs the 'mortality' process over an object,
decreasing the number of individuals. It is a generic, S3 methods can be specified
for a particular specification of the population.
}
\details{
The rate can be a single value or a value for each individual calculated
externally. No recycling is allowed.
}

