% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm-main.R
\docType{package}
\name{ibm-package}
\alias{ibm}
\alias{ibm-package}
\title{Individual Based Models in R}
\description{
Individual Based Models in R
}
\examples{
\dontrun{
} 
}
\author{
Ricardo Oliveros-Ramos Maintainer: Ricardo Oliveros-Ramos
<ricardo.oliveros@gmail.com>
}
\references{
ibm
}
\keyword{agent}
\keyword{based}
\keyword{ibm,}
\keyword{individual}
\keyword{models}
\keyword{models,}

